Fixed some syntax y stupid errors.
I'm codding this at the 2:00 am. I don't know what I'm doing.

+++ Editor/src/CustomImGuiWidgets.cpp
@@ -69,7 +69,7 @@ namespace rym {
 	}
 
 
-	bool DrawDragIntPlusMinus(const char* label, int* value, int reset, float min, float max)
+	bool DrawDragIntPlusMinus(const char* label, int* value, int reset, int min, int max)
 	{
 		ImGui::PushID(label);
 		ImGui::Text(label);
